src/main/webapp/app/entities/matricula/item-materia/route/item-materia-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IItemMateria, ItemMateria } from '../item-materia.model';
import { ItemMateriaService } from '../service/item-materia.service';

import { ItemMateriaRoutingResolveService } from './item-materia-routing-resolve.service';

describe('ItemMateria routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ItemMateriaRoutingResolveService;
  let service: ItemMateriaService;
  let resultItemMateria: IItemMateria | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(ItemMateriaRoutingResolveService);
    service = TestBed.inject(ItemMateriaService);
    resultItemMateria = undefined;
  });

  describe('resolve', () => {
    it('should return IItemMateria returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultItemMateria = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultItemMateria).toEqual({ id: 123 });
    });

    it('should return new IItemMateria if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultItemMateria = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultItemMateria).toEqual(new ItemMateria());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as ItemMateria })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultItemMateria = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultItemMateria).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
